Fix Float to Double type mismatch for topP parameter

+++ app/src/main/java/dev/aurakai/auraframefx/oracledrive/genesis/ai/clients/VertexAIClientImpl.kt
@@ -123,7 +123,7 @@ class VertexAIClientImpl @Inject constructor(
             ),
             generationConfig = GenerationConfig(
                 temperature = temperature.toDouble(),
-                topP = config.defaultTopP,
+                topP = config.defaultTopP.toDouble(),
                 topK = config.defaultTopK,
                 maxOutputTokens = maxTokens,
                 candidateCount = 1
